refactor: split lib, schemas, and sources modules
lib/ cleanup (L1-L3):
- Delete lib/{formatting,log,theme}.py; move to rendering/formatting.py,
  polylogue/logging.py, ui/theme.py respectively
- Add lib/filter_executor.py: _ExecutionPlan, build_execution_plan,
  sql_pushdown_params (was incorrectly in storage/, violating dep order)
- Add cli/filter_picker.py: interactive filter pick() UI

Schema splits (SC1):
- schemas/schema_inference.py: 1309→162 LOC thin re-export hub
- Add schemas/sampling.py: ProviderConfig, PROVIDERS, load_samples_*
- Add schemas/schema_generation.py: GenerationResult, collapse_dynamic_keys,
  generate_provider_schema, generate_all_schemas
- Add schemas/field_stats.py: FieldStats, _collect_field_stats, UUID_PATTERN
- Add schemas/privacy.py: _is_safe_enum_value, _is_content_field, denylist

Sources splits (SR1):
- Add sources/cursor.py: cursor state management
- Add sources/decoders.py: byte decoding helpers (_decode_json_bytes with BOM fix)
- Add sources/emitter.py: _ConversationEmitter, _ParseContext

Dependency violations fixed:
- lib/filters.py no longer imports from storage/ (filter_executor moved to lib/)
- schemas/{schema_inference,verification}.py now import db_path from
  polylogue.paths instead of storage/backends/connection
