+++ polylogue/sources/parsers/codex.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 from pydantic import ValidationError
 
 from polylogue.lib.branch_type import BranchType
-from polylogue.lib.log import get_logger
+from polylogue.logging import get_logger
 from polylogue.lib.roles import Role
 from polylogue.sources.providers.codex import CodexRecord
 from polylogue.types import Provider
